Moved sub classes to external file
